make cleanup command skip archived cbz files

+++ js/Crackers.js
@@ -319,12 +319,10 @@
     };
 
     Crackers.prototype.cleanup = function(args) {
-      var archive, archives, cmd, filename, j, len;
+      var archive, archivedCount, archives, archivesDir, cmd, filename, j, len, parsed, processedCount, rootDir;
       archives = this.findArchives(args.filenames);
-      if (archives.length === 0) {
-        log.warning("cleanup: Nothing to do!");
-        return;
-      }
+      archivedCount = 0;
+      processedCount = 0;
       cmd = "rm";
       if (process.platform === 'win32') {
         cmd = "del";
@@ -332,12 +330,29 @@
       for (j = 0, len = archives.length; j < len; j++) {
         archive = archives[j];
         filename = archive.abs;
+        parsed = path.parse(filename);
+        rootDir = cfs.findParentContainingFilename(filename, constants.ROOT_FILENAME);
+        if (!rootDir) {
+          log.warning("Skipping " + filename + ", not in a crackers root");
+          continue;
+        }
+        archivesDir = cfs.join(rootDir, constants.ARCHIVES_DIR);
+        if (cfs.insideDir(filename, archivesDir)) {
+          archivedCount += 1;
+          log.verbose("Skipping archived comic: " + filename);
+          continue;
+        }
         if (args.execute) {
           console.log("Removing: " + filename);
           fs.unlinkSync(filename);
         } else {
           console.log(cmd + " \"" + filename + "\"");
         }
+        processedCount += 1;
+      }
+      if (processedCount === 0) {
+        log.warning("cleanup: Nothing to do! (skipped " + archivedCount + " archived comics)");
+        return;
       }
     };
